LapChart - helpers, types, restructure
